+++ ariadne/asgi.py
@@ -44,6 +44,9 @@ GQL_COMPLETE = "complete"  # Server -> Client
 GQL_STOP = "stop"  # Client -> Server
 
 ExtensionList = Optional[List[Type[Extension]]]
+Extensions = Union[
+    Callable[[Any, Optional[ContextValue]], ExtensionList], ExtensionList
+]
 
 
 class GraphQL:
@@ -56,7 +59,7 @@ class GraphQL:
         debug: bool = False,
         logger: Optional[str] = None,
         error_formatter: ErrorFormatter = format_error,
-        extensions: Union[Callable[[Any], ExtensionList], ExtensionList] = None,
+        extensions: Optional[Extensions] = None,
         middleware: Optional[MiddlewareManager] = None,
         keepalive: float = None,
     ):
@@ -87,9 +90,14 @@ class GraphQL:
 
         return self.context_value or {"request": request}
 
-    async def get_extensions_for_request(self, request: Any) -> ExtensionList:
+    async def get_extensions_for_request(
+        self, request: Any, context: Optional[ContextValue]
+    ) -> ExtensionList:
         if callable(self.extensions):
-            return self.extensions(request)
+            extensions = self.extensions(request, context)
+            if isawaitable(extensions):
+                extensions = await extensions  # type: ignore
+            return extensions
         return self.extensions
 
     async def handle_http(self, scope: Scope, receive: Receive, send: Send):
@@ -118,7 +126,7 @@ class GraphQL:
             return PlainTextResponse(error.message or error.status, status_code=400)
 
         context_value = await self.get_context_for_request(request)
-        extensions = await self.get_extensions_for_request(request)
+        extensions = await self.get_extensions_for_request(request, context_value)
 
         success, response = await graphql(
             self.schema,